fineControl has been fixed
The function that let the controller be less sensitive at lower values
as been fixed and implemented into the code. This means should the
controller be moved up to a value of say 50, it will only power motor to
around 20, this way there is more precise control. However should the
controller be pushed up to its maximum of 127, it will still set motors
to 127 so no speed is sacrificed. The problem itself lay within the
floating point calculators of the C language, it somehow could not
implicitly convert 127 to 127.0 when set on a float variable.