Debugging Session
I went to test the code I commited to Github lately. Ultrasonic sensors
were replaced with working ones and autonomous code to go with it was
testing. Testing was successful. The robot would continue to move until
a condition was fufilled, in this case it was the ultrasonics. I also
tested the encoder condition and it seemed to be working. The only thing
giving me trouble now is the fine control, which apparently was not
fixed.

Using the ringtone converter, I also managed to upload a few tunes and
play them on the VEX speaker, so this might be interesting.