Emulation Debugging
RobotC compiler is not giving me any debug values for some reason, but
they are obviously there due to the fact that the robot still drives
when I apply the fineControl. Thus I have concluded the problem is due
to the float value being more than 0 but less than 1, which leads to a
rounding of 1. I have set the outputs from -3 to 3 to return 0.