Workshop Build Session
Today I managed to test several of the features last implemented in
emulation debugging. It seems the lift doesn't quite reach the bottom of
the preset I set, which is for minimum height. The fault was not in the
potentiometer, and was eventually found when I removed the 'easing'
code. This part will have to be redeveloped. It also seems that the
exponential drive doesn't work with the ramping code so I have scrapped
that completely now. It also seems the braking code for my autonomous
drive only accounted for driving forward and not backwards, I am working
on a fix for that.