Pre Workshop Commit
Changes to pretty much every piece of code file was required as I
updated the drive to not use seperate tasks anymore, as there were no
loops being introduced into the drive system as there looks to be none
in the future.

I fixed a problem with the autonomous not actually ending the step
properly, so the next action did not stop. A new time condition has been
added to add the option of waiting for a time condition to be met before
commands such as move were stopped. The safeguards previously written to
prevent conditions being conditions for the wrong steps were integrated
into the functions themselves so I don't have to manually write them
anymore.

Two new piston toggles added and remapped to the shoulder buttons on the
controller.

Lifttrim downwards only triggers at the start of a match, after this
there is no more downwards trim so that there's no slipping in driver
control or autonomous.

Lift compensation has been added in case the lift tilts when the skyrise
claw is added, this is not activated as of yet however.

Finally a bunch of new variables had to be declared to make these
functions work, stored within the variables header.