Autonomous Overhaul
Restored the old autonomous system with single conditional move
functions. New system too hard to debug with RobotC and the constant
hardware connection problems in VEX.

Added support for new buttons, one to control the needle deployment
without need for the controller and another to calibrate the light
sensor in different environments.

Autonomous currently should deploy the robot then drive forwards and
backwards to fill in skyrise pieces. It should detect whether there is a
skyrise or not in the holder with its light sensor and if so, grab it.
The robot deployment has enough force to knock the preload on to the
skyrise base, scoring additional points.