Workshop Session
Auton running into problems as older framework does not support multi
threaded actions, commands and conditions are tied together so may need
to rely on more tasks to get the job done.

The more advanced framework was proving too hard to debug, but other
than that it suited the current needs perfectly, might see if I can
restore it and fix.