Code update
This update brings about the fixes for the new auton framework, which is
fully working now due to the additional debugging functions I managed to
create. Because everything was so modular, I could just comment out
sections and add in code bit by bit, to find the offending function. I
found a few functions that were written incorrectly and corrected
accordingly.

Auton essentially runs as follows:
Picks up skyrise from holder, goes back into position, lift goes down
and drops skyrise into the skyrise base. Then it lifts up to the correct
height for loader before moving back again. A light sensor delays the
claw closing until a piece is detected in the holder, with a delay so
human error is accounted for.

Other additions with this patch includes a skyrise height preset toggle,
which will go up to each of the skyrise heights accordingly, and also an
autoloader height preset.